refactor(core): pen identifiers are int end to end (#44)
The archive column is int4 and the catalogue never answers anything
else; the long widening and the range-tested narrowing in the provider
were self-inflicted.

+++ SemiPlot/SemiPlot.Core/Trends/Pen.cs
@@ -1,7 +1,7 @@
 ﻿namespace SemiPlot.Core.Trends;
 
 public sealed record Pen(
-	long PenId,
+	int PenId,
 	string Name,
 	string Group,
 	string Color,
