+++ SemiPlot/SemiPlot.UI/Chart/HistoryRequest.cs
@@ -5,7 +5,7 @@ namespace SemiPlot.UI.Chart;
 // Sequence is a monotonic stamp assigned at request time so a slow query can never overwrite a newer
 // window's result.
 public sealed record HistoryRequest(
-	IReadOnlyList<long> PenIds,
+	IReadOnlyList<int> PenIds,
 	DateTime FromUtc,
 	DateTime ToUtc,
 	AggregationLayer Layer,
